+++ public/js/functions.js
@@ -13,6 +13,13 @@ function toggleSettings() {
   buttonPanel.classList.toggle("wn-settings-btn-active");
 }
 
+function toggleAbout() {
+  var elementPanel = document.getElementById("wn-about-panel");
+  elementPanel.classList.toggle("wn-about-hidden");
+  var buttonPanel = document.getElementById("wn-about-btn");
+  buttonPanel.classList.toggle("wn-about-btn-active");
+}
+
 function toggleCustomInput(id, selected) {
   var element = document.getElementById(id);
   if (selected === "custom") {
@@ -88,10 +95,16 @@ function loadFeed(url, idx) {
 }
 
 function updatePageTitle() {
-    let title = 'WhoNews Beta ';
-    const labels = document.querySelectorAll('.wn-tab-label');
-    [...labels].map((elt) => {
-       title += ' | ' + elt.innerHTML;
-    });
+    let title = 'WhoNews.org';
+
+    if (defaultPageTitle.length) {
+        title = defaultPageTitle;
+    } else {
+        const labels = document.querySelectorAll('.wn-tab-label');
+        [...labels].map((elt) => {
+           title += ' | ' + elt.innerHTML;
+        });
+    }
+
     document.querySelector('title').innerHTML = title;
 }